+++ node/src/types/kernel/api/filetree/getHPathByPath/index.d.ts
@@ -17,12 +17,7 @@
 
 
 /* directories */
-export * as createDocWithMd from "./createDocWithMd";
-export * as getHPathByPath from "./getHPathByPath";
-export * as removeDoc from "./removeDoc";
-export * as renameDoc from "./renameDoc";
 
 /* files */
-export * as listDocsByPath from "./listDocsByPath";
-export * as renameDoc from "./renameDoc";
-export * as searchDocs from "./searchDocs";
+export * from "./payload";
+export * from "./response";